savegame/common: port Savegame_ApplyLogicToCurrentInfo to TRX
