+++ src/libtrx/game/savegame/common.c
@@ -426,6 +426,122 @@ void Savegame_PersistGameToCurrentInfo(const GF_LEVEL *const level)
     }
 }
 
+void Savegame_ApplyLogicToCurrentInfo(const GF_LEVEL *const level)
+{
+    RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
+    if (resume == nullptr) {
+        return;
+    }
+
+    LOG_INFO("Applying game logic to level #%d", level->num);
+
+    if (!g_Config.gameplay.disable_healing_between_levels
+        || level == GF_GetGymLevel() || level == GF_GetFirstLevel()) {
+        resume->lara_hitpoints = g_Config.gameplay.start_lara_hitpoints;
+    }
+
+    if (level == GF_GetGymLevel()) {
+        resume->flags.available = true;
+        resume->flags.costume = TR_VERSION == 1;
+
+        resume->flags.has_pistols = false;
+        resume->flags.has_shotgun = false;
+        resume->flags.has_magnums = false;
+        resume->flags.has_uzis = false;
+
+        resume->small_medipacks = 0;
+        resume->large_medipacks = 0;
+        resume->pistol_ammo = 0;
+        resume->shotgun_ammo = 0;
+        resume->magnum_ammo = 0;
+        resume->uzi_ammo = 0;
+#if TR_VERSION == 1
+        resume->num_scions = 0;
+#else
+        resume->flags.has_harpoon = false;
+        resume->flags.has_m16 = false;
+        resume->flags.has_grenade = false;
+        resume->harpoon_ammo = 0;
+        resume->m16_ammo = 0;
+        resume->grenade_ammo = 0;
+        resume->flares = 0;
+#endif
+        resume->equipped_gun_type = LGT_UNARMED;
+        resume->holsters_gun_type = LGT_UNARMED;
+        resume->back_gun_type = LGT_UNARMED;
+        resume->gun_status = LGS_ARMLESS;
+    }
+
+    if (level == GF_GetFirstLevel()) {
+        resume->flags.available = true;
+        resume->flags.costume = false;
+
+        resume->flags.has_pistols = true;
+        resume->flags.has_shotgun = false;
+        resume->flags.has_magnums = false;
+        resume->flags.has_uzis = false;
+
+        resume->small_medipacks = 0;
+        resume->large_medipacks = 0;
+        resume->pistol_ammo = 1000;
+        resume->shotgun_ammo = 0;
+        resume->magnum_ammo = 0;
+        resume->uzi_ammo = 0;
+#if TR_VERSION == 1
+        resume->num_scions = 0;
+#else
+        resume->flags.has_harpoon = false;
+        resume->flags.has_m16 = false;
+        resume->flags.has_grenade = false;
+        resume->harpoon_ammo = 0;
+        resume->m16_ammo = 0;
+        resume->grenade_ammo = 0;
+        resume->flares = 0;
+#endif
+        resume->equipped_gun_type = LGT_PISTOLS;
+        resume->holsters_gun_type = LGT_PISTOLS;
+        resume->back_gun_type = LGT_UNARMED;
+        resume->gun_status = LGS_ARMLESS;
+    }
+
+    if (Game_IsBonusFlagSet(GBF_NGPLUS) && level != GF_GetGymLevel()) {
+        resume->flags.has_pistols = true;
+        resume->flags.has_shotgun = true;
+        resume->flags.has_magnums = true;
+        resume->flags.has_uzis = true;
+
+        resume->shotgun_ammo = 10000;
+        resume->magnum_ammo = 10000;
+        resume->uzi_ammo = 10000;
+
+#if TR_VERSION == 1
+        resume->equipped_gun_type = LGT_UZIS;
+        resume->holsters_gun_type = LGT_UZIS;
+        resume->back_gun_type = LGT_SHOTGUN;
+#else
+        resume->flags.has_m16 = true;
+        resume->flags.has_grenade = true;
+        resume->flags.has_harpoon = true;
+        resume->m16_ammo = 10000;
+        resume->grenade_ammo = 10000;
+        resume->harpoon_ammo = 10000;
+
+        resume->flares = -1;
+        resume->equipped_gun_type = LGT_GRENADE;
+        resume->holsters_gun_type = LGT_PISTOLS;
+        resume->back_gun_type = LGT_GRENADE;
+#endif
+    }
+
+#if TR_VERSION == 2
+    const STATS_COMMON default_stats = Savegame_GetDefaultStats(level);
+    resume->stats.max_secret_count = default_stats.max_secret_count;
+    resume->stats.all_secrets_mask = default_stats.all_secrets_mask;
+#endif
+
+    Savegame_DetermineLegacyGunTypes(resume);
+}
+
 void Savegame_ProcessItemsBeforeSave(void)
 {
     for (int32_t i = 0; i < Item_GetLevelCount(); i++) {
